+++ lib/glific/clients/common_webhook.ex
@@ -3,8 +3,13 @@ defmodule Glific.Clients.CommonWebhook do
   Common webhooks which we can call with any clients.
   """
 
-  alias Glific.OpenAI.ChatGPT
-  alias Glific.Sheets.GoogleSheets
+  alias Glific.{
+    ASR.GoogleASR,
+    Contacts.Contact,
+    OpenAI.ChatGPT,
+    Repo,
+    Sheets.GoogleSheets
+  }
 
   @doc """
   Create a webhook with different signatures, so we can easily implement
@@ -67,6 +72,22 @@ defmodule Glific.Clients.CommonWebhook do
 
       {_status, _response} ->
         %{success: false, response: "Invalid response"}
+    end
+  end
+
+  # This webhook will call Google speech-to-text API
+  def webhook("speech_to_text", fields) do
+    contact_id = Glific.parse_maybe_integer!(fields["contact"]["id"])
+    contact = get_contact_language(contact_id)
+
+    Glific.parse_maybe_integer!(fields["organization_id"])
+    |> GoogleASR.speech_to_text(fields["results"], contact.language.locale)
+  end
+
+  defp get_contact_language(contact_id) do
+    case Repo.fetch(Contact, contact_id) do
+      {:ok, contact} -> contact |> Repo.preload(:language)
+      {:error, error} -> error
     end
   end
